% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cdf_cpp}
\alias{cdf_cpp}
\title{CDF of GLM}
\usage{
cdf_cpp(y, X, beta, phi, distname, linkname, n)
}
\arguments{
\item{y}{response \code{vector}}

\item{X}{design \code{matrix}}

\item{beta}{regression coefficient \code{vector}}

\item{phi}{Dispersion parameter. Ignored for binomial and Poisson models}

\item{distname}{name of distribution as a string. Must be one of \code{ c ( "gaussian", "gamma", "poisson", "binomial" ) ) }}

\item{linkname}{string giving name of link function. Must be one of \code{ c( "logit", "probit", "cauchit", "cloglog", "identity", "log", "sqrt", "1/mu^2", "inverse" ) }}

\item{n}{number of observations}
}
\value{
vector applying CDF to \eqn{ y \mid X, \beta, \phi }
}
\description{
This function computes the CDF for each observation y in a GLM
}
