% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{conv_to_normal}
\alias{conv_to_normal}
\title{Convert to Normal}
\usage{
conv_to_normal(y, X, beta, phi, distname, linkname, n)
}
\arguments{
\item{y}{response \code{vector}}

\item{X}{design \code{matrix}}

\item{beta}{regression coefficient \code{vector}}

\item{phi}{Dispersion parameter. Ignored for binomial and Poisson models}

\item{distname}{name of distribution as a string. Must be one of \code{ c ( "gaussian", "Gamma", "poisson", "binomial" ) ) }}

\item{linkname}{string giving name of link function. Must be one of \code{ c( "logit", "probit", "cauchit", "cloglog", "identity", "log", "sqrt", "1/mu^2", "inverse" ) }}

\item{n}{number of observations}
}
\value{
vector of standard normal variables
}
\description{
This function computes \eqn{h_{ij}^{-1}(y_{ij}) = \Phi^{-1}(F_{ij}(y_{ij} | \beta_j, \phi_j))}
}
