% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{copula_predict}
\alias{copula_predict}
\title{Posterior predictive sample of copula GLM}
\usage{
copula_predict(
  Xlist,
  distnamevec,
  linknamevec,
  Gamma,
  betasample,
  phisample,
  n,
  J
)
}
\arguments{
\item{Xlist}{a \code{list} of length \eqn{J}, each element is a design \code{matrix}}

\item{distnamevec}{a \code{character} vector of length \eqn{J} giving the name of the distribution of each endpoint}

\item{linknamevec}{a \code{character} vector of length \eqn{J} giving the name of the link function of each endpoint}

\item{Gamma}{a sampled correlation \code{matrix}}

\item{betasample}{a \eqn{J}-dimensional \code{list} of sampled regression coefficients}

\item{phisample}{a \code{vector} of sampled dispersion parameters}

\item{n}{sample size for future data}

\item{J}{number of endpoints}
}
\value{
a \eqn{n \times J} matrix of samples from the predictive posterior density
}
\description{
Obtain a sample from the posterior predictive density of a copula GLM
}
