% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{copula_predict_all}
\alias{copula_predict_all}
\title{List of posterior predictive samples of copula GLM}
\usage{
copula_predict_all(
  Xlistlist,
  distnamevec,
  linknamevec,
  betasamplelist,
  phisamplemat,
  Gammaarray,
  n,
  J,
  M
)
}
\arguments{
\item{Xlistlist}{a \code{M}-dimensional \code{list} of lists. The inner list is a list of design matrices of length \code{J}}

\item{distnamevec}{a \code{character} vector of length \eqn{J} giving the name of the distribution of each endpoint}

\item{linknamevec}{a \code{character} vector of length \eqn{J} giving the name of the link function of each endpoint}

\item{betasamplelist}{a list of length \code{M}. Each element is a list of length \code{J} giving the design matrix for each endpoint}

\item{phisamplemat}{a \eqn{M \times J} matrix of sampled dispersion parameters}

\item{Gammaarray}{a \eqn{J \times J \times M} array of sampled correlation matrices}

\item{n}{sample size for future data}

\item{J}{number of endpoints}

\item{M}{number of samples}
}
\value{
a \eqn{n \times J} matrix of samples from the predictive posterior density
}
\description{
Obtain a sample from the posterior predictive density of a copula GLM
}
