% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmPOS-package.R
\docType{package}
\name{glmPOS-package}
\alias{glmPOS}
\alias{glmPOS-package}
\title{glmPOS: What the Package Does Using Title Case}
\description{
More details about what the package does. See
             <http://cran.r-project.org/doc/manuals/r-release/R-exts.html#The-DESCRIPTION-file> for details on how
             to write this part.
}
\keyword{internal}
