% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{invcdf_cpp}
\alias{invcdf_cpp}
\title{Inverse CDF of GLM}
\usage{
invcdf_cpp(u, X, beta, phi, distname, linkname, n)
}
\arguments{
\item{u}{a \code{vector} consisting of values between 0 and 1}

\item{X}{design \code{matrix}}

\item{beta}{regression coefficient \code{vector}}

\item{phi}{Dispersion parameter. Ignored for binomial and Poisson models}

\item{distname}{name of distribution as a string. Must be one of \code{ c ( "gaussian", "gamma", "poisson", "binomial" ) ) }}

\item{linkname}{string giving name of link function. Must be one of \code{ c( "logit", "probit", "cauchit", "cloglog", "identity", "log", "sqrt", "1/mu^2", "inverse" ) }}

\item{n}{number of observations}
}
\value{
vector applying CDF to \eqn{ y \mid X, \beta, \phi }
}
\description{
This function computes the Inverse CDF for each observation y in a GLM
}
