% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{linkinv_cpp}
\alias{linkinv_cpp}
\title{Inverse link}
\usage{
linkinv_cpp(eta, linkname)
}
\arguments{
\item{eta}{\code{vector} of linear predictors}

\item{linkname}{string giving name of link function. Must be one of
\code{ c( "logit", "probit", "cauchit", "cloglog", "identity", "log", "sqrt", "1/mu^2", "inverse" ) }}
}
\value{
\code{vector} applying inverse link function to \code{eta}
}
\description{
This function takes as input a linear predictor eta and the name of a link function,
and outputs the inverse link function applied to eta
}
