% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logInitPrior_cpp}
\alias{logInitPrior_cpp}
\title{Initial validation prior for (beta, phi)}
\usage{
logInitPrior_cpp(beta, phi, c0, alpha0, gamma0, p)
}
\arguments{
\item{beta}{\code{vector} of regression coefficients}

\item{phi}{dispersion parameter}

\item{c0}{prior variance multiple for regression coefficients given inverse dispersion}

\item{alpha0}{shape parameter for inverse dispersion}

\item{gamma0}{rate parameter for inverse dispersion}

\item{p}{dimension of \code{beta}}
}
\value{
scalar giving log prior joint density
}
\description{
This function returns the log prior density for (beta, phi) where
\eqn{\beta \mid \phi \sim N_p(0, \phi c_0 I)} and
\eqn{\phi \sim } Gamma \eqn{ ( \alpha_0, \gamma_0 ) }
}
