% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglik_cpp}
\alias{loglik_cpp}
\title{Log likelihood of GLM}
\usage{
loglik_cpp(y, X, beta, phi, distname, linkname, n)
}
\arguments{
\item{y}{response \code{vector}}

\item{X}{design \code{matrix}}

\item{beta}{regression coefficient \code{vector}}

\item{phi}{Dispersion parameter. Ignored for binomial and Poisson models}

\item{distname}{name of distribution as a string. Must be one of \code{ c ( "gaussian", "Gamma", "poisson", "binomial" ) ) }}

\item{linkname}{string giving name of link function. Must be one of \code{ c( "logit", "probit", "cauchit", "cloglog", "identity", "log", "sqrt", "1/mu^2", "inverse" ) }}

\item{n}{number of observations}
}
\value{
scalar giving log likelihood
}
\description{
This function computes log likelihood based on data, parameters, link function, and a string giving
the proper distribution
}
