% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logPost}
\alias{logPost}
\title{Log joint posterior density for (beta, phi)}
\usage{
logPost(
  y,
  X,
  beta,
  phi,
  Z,
  Gammainv,
  distname,
  linkname,
  n,
  j,
  J,
  p,
  c0,
  alpha0,
  gamma0,
  b0,
  y0,
  X0,
  n0
)
}
\arguments{
\item{y}{response \code{vector} for current data}

\item{X}{design \code{matrix} for current data}

\item{beta}{\code{vector} of regression coefficients}

\item{phi}{dispersion parameter}

\item{Z}{\code{matrix} giving unobserved variables}

\item{Gammainv}{inverse correlation \code{matrix}}

\item{distname}{\code{character} giving which distribution to use}

\item{linkname}{\code{character} giving which link function to use}

\item{n}{sample size of current data}

\item{j}{index of which margin to compute likelihood for, \eqn{0 \le j \le J - 1}}

\item{J}{number of endpoints of current data,}

\item{p}{number of covariates in current endpoint,}

\item{c0}{variance multiple in \eqn{\beta \mid \phi} normal prior}

\item{alpha0}{Gamma distribution shape parameter for prior on \code{phi}}

\item{gamma0}{Gamma distribution rate parameter for prior on \code{phi}}

\item{b0}{power prior parameter for historical data}

\item{y0}{\code{vector} of historical responses}

\item{X0}{design \code{matrix} for historical data}

\item{n0}{historical data sample size}
}
\value{
scalar giving log posterior density for (beta, phi)
}
\description{
This function returns the log joint posterior density \eqn{(\beta, \phi)}
}
