% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logPowerPrior_cpp}
\alias{logPowerPrior_cpp}
\title{Power prior for \eqn{(\beta, \phi)}}
\usage{
logPowerPrior_cpp(y0, X0, beta, phi, b0, distname, linkname, n0)
}
\arguments{
\item{y0}{\code{vector} of historical responses}

\item{X0}{design \code{matrix} for historical data}

\item{beta}{regression coefficients}

\item{phi}{dispersion parameter}

\item{b0}{power prior parameter, a \code{b0} \eqn{\in (0, 1])}}

\item{distname}{name of distribution}

\item{linkname}{name of link function}

\item{n0}{historical data sample size}
}
\value{
scalar giving log prior density
}
\description{
This function returns the log power prior for \eqn{(\beta, \phi)}
}
