% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_copula.R
\name{sample_copula}
\alias{sample_copula}
\title{Sample from Bayesian copula GLM}
\usage{
sample_copula(
  formula.list,
  family.list,
  data,
  M = 10000,
  data0 = NULL,
  b0 = NULL,
  c0 = NULL,
  alpha0 = NULL,
  gamma0 = NULL,
  Gamma = NULL,
  S0beta = NULL,
  sigma0logphi = NULL,
  v0 = NULL,
  V0 = NULL,
  beta0 = NULL,
  phi0 = NULL
)
}
\arguments{
\item{formula.list}{A \eqn{J}-dimensional list of formulas giving how the endpoints are related to the covariates}

\item{family.list}{A \eqn{J}-dimensional list of families giving how each endpoint is distributed. See \code{help(family)}}

\item{data}{A \code{data frame} containing all response variables and covariates. Variables must be named.}

\item{M}{Number of desired posterior samples}

\item{data0}{\emph{Optional} historical data set for power prior on \eqn{\beta, \phi}}

\item{b0}{\emph{Optional} power prior hyperparameter}

\item{c0}{A \eqn{J}-dimensional vector for \eqn{\beta \mid \phi} prior covvariance: \eqn{\text{Cov}(\beta \mid \phi) = c_0 \phi I_n }. If \code{NULL}, sets \eqn{c_0 = 10000} for each endpoint}

\item{alpha0}{A \eqn{J}-dimensional vector giving the shape hyperparameter for each dispersion parameter on the prior on \eqn{\phi}. If \code{NULL} sets \eqn{\alpha_0 = .01} for each dispersion parameter}

\item{gamma0}{A \eqn{J}-dimensional vector giving the rate hyperparameter for each dispersion parameter on the prior on \eqn{\phi}. If \code{NULL} sets \eqn{\alpha_0 = .01} for each dispersion parameter}

\item{Gamma}{Initial value for correlation matrix. If \code{NULL} defaults to the correlation matrix from the responses.}

\item{S0beta}{A \eqn{J}-dimensional \code{list} for the covariance matrix for random walk metropolis on beta. Each matrix must have the same dimension as the corresponding regression coefficient. If \code{NULL}, uses \code{solve(crossprod(X))}}

\item{sigma0logphi}{A \eqn{J}-dimensional \code{vector} giving the standard deviation on \eqn{\log(\phi)} for random walk metropolis. If \code{NULL} defaults to \code{0.1}}

\item{v0}{An integer scalar giving degrees of freedom for Inverse Wishart prior. If \code{NULL} defaults to J + 2}

\item{V0}{An integer giving inverse scale parameter for Inverse Wishart prior. If \code{NULL} defaults to \code{diag(.001, J)}}

\item{beta0}{A \eqn{J}-dimensional \code{list} giving starting values for random walk Metropolis on the regression coefficients. If \code{NULL}, defaults to the GLM MLE}

\item{phi0}{A \eqn{J}-dimensional \code{vector} giving initial values for dispersion parameters. If \code{NULL}. Dispersion parameters will alwyas return 1 for binomial and Poisson models.}
}
\value{
A named list. 
\code{["betasample"]} gives a \eqn{J}-dimensional list of sampled coefficients as matrices. 
\code{["phisample"]} gives a \eqn{M \times J} matrix of sampled dispersion parameters. 
\code{["Gammasample"]} gives a \eqn{J \times J \times M} array of sampled correlation matrices. 
\code{["betaaccept"]} gives a \eqn{M \times J} matrix where each row indicates whether the proposal for the regression coefficient was accepted.
\code{["phiaccept"]} gives a \eqn{M \times J} matrix where each row indicates whether the proposal for the dispersion parameter was accepted
}
\description{
Sample from any GLM via Bayesian copula regression as developed in Pitt et al. (2006).
Uses random-walk Metropolis to update regression coefficients and dispersion parameters.
Assumes Inverse Wishart prior on augmented data with covariance matrix as developed by Hoff (2007).
}
