% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_copula_cpp}
\alias{sample_copula_cpp}
\title{MCMC sample of copula GLM}
\usage{
sample_copula_cpp(
  ymat,
  Xlist,
  distnamevec,
  linknamevec,
  c0vec,
  S0betalist,
  sigma0logphivec,
  alpha0vec,
  gamma0vec,
  Gamma,
  v0,
  V0,
  b0,
  y0mat,
  X0list,
  M,
  beta0list,
  phi0vec
)
}
\arguments{
\item{ymat}{\eqn{n \times J} \code{matrix} of response variables}

\item{Xlist}{\eqn{J}-dimensional list of design matrices}

\item{distnamevec}{\code{character} vector of length \eqn{J}  giving name of distribution}

\item{linknamevec}{\code{character} vector of length \eqn{J}  giving name of link function. See \code{help(family)}}

\item{c0vec}{\code{numeric} vector giving scale hyperparmeter for conditional prior. The covariance is c0 * phi * S0beta.}

\item{S0betalist}{\code{list} of matrices; each matrix gives prior covariance: the prior covariance is c0 * phi * S0beta}

\item{sigma0logphivec}{\code{vector} giving random walk variance for log dispersion parameter}

\item{alpha0vec}{\code{numeric} vector giving shape parameter for the gamma density prior on phi}

\item{gamma0vec}{\code{numeric} vector giving rate parameter for the gamma density prior on phi}

\item{Gamma}{\code{matrix} giving starting values for correlation parameters}

\item{v0}{integer giving degrees of freedom for Inverse wishart prior on correlation matrix \code{Gamma}}

\item{V0}{\code{matrix} giving prior scale parameter for Inverse wishart prior on correlation matrix \code{Gamma}}

\item{b0}{scalar between 0 and 1 giving power prior hyperparameter. If \code{b0 == 0}, historical data is ignored.}

\item{y0mat}{\code{matrix} giving responses for historical data set}

\item{X0list}{\code{list} giving design matrices for historical data responses}

\item{M}{number of samples to draw}

\item{beta0list}{\code{list} of vectors giving starting values for regression coefficients}

\item{phi0vec}{\code{numeric} vector giving starting values for dispersion parameters}
}
\value{
sampled correlation matrix
}
\description{
This function samples from the posterior distribution of a copula GLM
}
