% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_y}
\alias{sample_y}
\title{Sample GLM copula response variable}
\usage{
sample_y(z, X, beta, phi, distname, linkname, n)
}
\arguments{
\item{z}{\code{vector} of \eqn{N(0, 1)} samples from \eqn{Z \sim N(0, \Gamma)}}

\item{X}{design \code{matrix}}

\item{beta}{regression coefficient \code{vector}}

\item{phi}{Dispersion parameter. Ignored for binomial and Poisson models}

\item{distname}{name of distribution as a string. Must be one of \code{ c ( "gaussian", "gamma", "poisson", "binomial" ) ) }}

\item{linkname}{string giving name of link function. Must be one of \code{ c( "logit", "probit", "cauchit", "cloglog", "identity", "log", "sqrt", "1/mu^2", "inverse" ) }}

\item{n}{number of observations}
}
\value{
\code{vector} applying inverse CDF
}
\description{
This function computes the response variable y from a copula GLM given Z, a sample from N(0, Gamma)
}
