% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_Gamma}
\alias{update_Gamma}
\title{Update correlation matrix}
\usage{
update_Gamma(Z, n, v0, v0V0)
}
\arguments{
\item{Z}{a \eqn{n \times J} \code{matrix} of latent variables}

\item{n}{number of rows in Z}

\item{v0}{hyperparameter giving degrees of freedom for inverse Wishart sample}

\item{v0V0}{v0 * hyperparameter giving scale matrix for inverse Wishart sample}
}
\value{
sampled correlation matrix
}
\description{
This function samples from the posterior correlation matrix given
the latent variables given as a matrix Z
}
