% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_Z}
\alias{update_Z}
\title{Get updated Z}
\usage{
update_Z(y, X, beta, phi, Z, Gamma, distname, linkname, n, j)
}
\arguments{
\item{y}{response \code{vector}}

\item{X}{design \code{matrix}}

\item{beta}{regression coefficient \code{vector}}

\item{phi}{Dispersion parameter. Ignored for binomial and Poisson models}

\item{Z}{current \code{matrix} of latent variables}

\item{Gamma}{current correlation \code{matrix}}

\item{distname}{name of distribution as a string. Must be one of \code{ c ( "gaussian", "Gamma", "poisson", "binomial" ) ) }}

\item{linkname}{string giving name of link function. Must be one of \code{ c( "logit", "probit", "cauchit", "cloglog", "identity", "log", "sqrt", "1/mu^2", "inverse" ) }}

\item{n}{number of observations}

\item{j}{index of which column of Z to update, an integer between 0 and J-1}
}
\value{
vector applying CDF to \eqn{ y \mid X, \beta, \phi }
}
\description{
This function updates the hidden variables in a GLM copula based on \eqn{\beta, \phi}
}
