% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayescopulareg-package.R
\docType{package}
\name{bayescopulareg-package}
\alias{bayescopulareg}
\alias{bayescopulareg-package}
\title{bayescopulareg: Bayesian Copula Regression}
\description{
Tools for Bayesian copula generalized linear models (GLMs). 
             The sampling scheme is based on Pitt, Chan, and Kohn (2006) <doi:10.1093/biomet/93.3.537>. 
             Regression parameters (including coefficients and dispersion parameters) are
             estimated via the adaptive random walk Metropolis approach developed by
             Haario, Saksman, and Tamminen (1999) <doi:10.1007/s001800050022>.
             The prior for the correlation matrix is based on Hoff (2007) <doi:10.1214/07-AOAS107>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ethan-alt/bayescopulareg}
  \item Report bugs at \url{https://github.com/ethan-alt/bayescopulareg/issues}
}

}
\keyword{internal}
