% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayescopulareg-package.R
\docType{package}
\name{bayescopulareg-package}
\alias{bayescopulareg}
\alias{bayescopulareg-package}
\title{bayescopulareg: Bayesian Copula Regression}
\description{
Tools for Bayesian copula generalized linear models (GLMs). 
             Includes posterior sampling and predictive posterior sampling. 
            Methods based on Pitt, Chan, and Kohn (2006) <doi:10.1093/biomet/93.3.537>. 
            Prior for correlation matrix based on Hoff (2007) <doi:10.1214/07-AOAS107>.
}
\keyword{internal}
